Fix: Sustain modifiers throught the note
Before, modifiers (i.e., Octave) were applied live to all the notes.
If a modifier was changed while a note was releasing (decaying),
it sounded very weird.

Now, modifiers are applied when a new note is hit, and the applied
modifiers to a note sustain throughout its lifetime, especially
while decaying.

As a side effect, modifiers are now much more intuitive.